Added checks around calls to vtkgl::DeleteShader to make sure the function pointers for vtkgl::DeleteShader have been loaded prior to calling them.

+++ lib/VTK/Rendering/vtkGatherFluorescencePolyDataMapper.cxx
@@ -72,14 +72,16 @@ void vtkGatherFluorescencePolyDataMapper::ReleaseGraphicsResources(vtkWindow *wi
   }
   this->LastWindow = NULL;
 
-  vtkgl::DeleteShader(this->VertexProgramHandle);
-  this->VertexProgramHandle = 0;
+  if (vtkgl::DeleteShader) {
+    vtkgl::DeleteShader(this->VertexProgramHandle);
+    this->VertexProgramHandle = 0;
 
-  vtkgl::DeleteShader(this->FragmentProgramHandle);
-  this->FragmentProgramHandle = 0;
+    vtkgl::DeleteShader(this->FragmentProgramHandle);
+    this->FragmentProgramHandle = 0;
 
-  vtkgl::DeleteProgram(this->ShaderProgramHandle);
-  this->ShaderProgramHandle = 0;
+    vtkgl::DeleteProgram(this->ShaderProgramHandle);
+    this->ShaderProgramHandle = 0;
+  }
 }
